pyproject.toml
[project]
name = "podcast-to-text"
version = "0.1.0"
description = "Convert podcast audio to formatted Markdown text"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp",
    "azure-cognitiveservices-speech",
    "openai",
    "pydub",
    "audioop-lts",
    "python-dotenv",
    "click",
    "rich",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/podcast_to_text"]
